pyproject.toml
[project]
name = "pa"
version = "0.1.0"
description = "Personal financial assistant"
requires-python = ">=3.11"
dependencies = [
    "argon2-cffi>=23.1.0",
    "cryptography>=42.0.0",
    "aiosqlite>=0.20.0",
    # sqlcipher3-binary is required on the Raspberry Pi target; omitted here as it does not build on Windows
    "playwright>=1.40.0",
    "python-telegram-bot>=20.0",
    "apscheduler>=4.0.0a5",
    "anthropic>=0.40.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"
